# Conflicts:
#	src/main/resources/templates/kaohe/all_score.html

+++ src/main/java/com/coolwen/experimentplatform/controller/ExpModelController.java
@@ -16,6 +16,9 @@ import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpSession;
+import java.text.ParseException;
+import java.text.SimpleDateFormat;
+import java.util.Date;
 import java.util.List;
 
 
@@ -51,6 +54,10 @@ public class ExpModelController {
     TotalScoreCurrentService totalScoreCurrentService;
     @Autowired
     ClazzService clazzService;
+    @Autowired
+    CollegeReportService collegeReportService;
+    @Autowired
+    DockerService dockerService;
 
 
     //查询模块信息页面
@@ -73,6 +80,7 @@ public class ExpModelController {
                       int m_classhour,
                       String m_inurl,
                       MultipartFile m_image,
+                      boolean report_type,
                       HttpServletRequest request,
                       HttpSession session
                       )
@@ -85,6 +93,7 @@ public class ExpModelController {
         expModel.setClasshour(m_classhour);
         expModel.setM_inurl(m_inurl);
         expModel.setImageurl(request.getScheme() + "://" + request.getServerName() + ":" + request.getServerPort()+"/ExperimentPlatform/ExpModelImage/"+file);
+        expModel.setReport_type(report_type);
         expModelService.save(expModel);
         session.setAttribute("modelId",expModel.getM_id());
         return "redirect:/expmodel/addTheory";
@@ -120,8 +129,14 @@ public class ExpModelController {
         moduleTestQuestService.deleteAllModuleTestQuest(moduleTestQuests);
         //删除实验模块报告与实验报告回答
         List<Report> reports = reportService.findReportByMId(id);
-        for(Report r : reports){
-            reportAnswerService.deleteReportAnswerByReportId(r.getReportId());
+        if(reports != null && !reports.isEmpty()){
+            for(Report r : reports){
+                reportAnswerService.deleteReportAnswerByReportId(r.getReportId());
+            }
+        }
+        List<CollegeReport> collegeReportList = collegeReportService.findCollegeReportByMid(id);
+        if(collegeReportList != null && !collegeReportList.isEmpty()){
+            collegeReportService.deleteCollegeList(collegeReportList);
         }
         reportService.deleteReports(reports);
         expModelService.deleteExpModelById(id);
@@ -144,6 +159,7 @@ public class ExpModelController {
                                   int m_classhour,
                                   String m_inurl,
                                   MultipartFile m_image,
+                                  boolean report_type,
                                   HttpServletRequest request,
                                   @PathVariable("id") int id
                                   )
@@ -154,9 +170,19 @@ public class ExpModelController {
         preExpModel.setM_type(m_type);
         preExpModel.setClasshour(m_classhour);
         preExpModel.setM_inurl(m_inurl);
+        preExpModel.setReport_type(report_type);
         String path = fIleService.upload(request,m_image);
         if(path != null){
             preExpModel.setImageurl(request.getScheme() + "://" + request.getServerName() + ":" + request.getServerPort()+"/ExperimentPlatform/ExpModelImage/"+path);
+        }
+        if(report_type == false){
+            //修改为自定义版删学院版答题记录
+            collegeReportService.deleteCollege(id);
+        }else {
+            List<Report> reportList = reportService.findReportByMId(id);
+            for (Report r : reportList){
+                reportAnswerService.deleteReportAnswerByReportId(r.getReportId());
+            }
         }
         expModelService.save(preExpModel);
         return "redirect:/expmodel/list";
@@ -299,22 +325,73 @@ public class ExpModelController {
 
     //实验大厅所有模块信息
     @GetMapping("/alltestModel")
-    public String alltest(Model model,@RequestParam(value = "pageNum",required = true,defaultValue = "0")int pageNum,HttpSession session){
-        Page<ExpModel> page = expModelService.finExpAll(pageNum);
+    public String alltest(Model model,@RequestParam(value = "pageNum",required = true,defaultValue = "0")int pageNum,HttpSession session) throws ParseException {
+        SimpleDateFormat simpleDateFormat = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss");
         model.addAttribute("list",expModelService.finExpAll(pageNum));
         session.setAttribute("modulePageNum",pageNum);
         session.setAttribute("isAllModule",true);
+        Student student = (Student) SecurityUtils.getSubject().getPrincipal();
+        Docker docker = dockerService.findDockerByStu_id(student.getId());
+        long nowDate = new Date().getTime();
+        String flag = "1314-06-21 00:00:00";
+        if(docker != null){
+            if(simpleDateFormat.format(docker.getDc_start_datetime()).equals(flag) && simpleDateFormat.format(docker.getDc_end_datetime()).equals(flag)){
+                model.addAttribute("docker",docker);
+                return "home_shiyan/all-test";
+            }else if(!simpleDateFormat.format(docker.getDc_start_datetime()).equals(flag) && simpleDateFormat.format(docker.getDc_end_datetime()).equals(flag)){
+                if(nowDate > docker.getDc_start_datetime().getTime()){
+                    model.addAttribute("docker",docker);
+                    return "home_shiyan/all-test";
+                }
+            }else if(simpleDateFormat.format(docker.getDc_start_datetime()).equals(flag) && !simpleDateFormat.format(docker.getDc_end_datetime()).equals(flag)){
+                if(nowDate < docker.getDc_end_datetime().getTime()){
+                    model.addAttribute("docker",docker);
+                    return "home_shiyan/all-test";
+                }
+            } else if(!simpleDateFormat.format(docker.getDc_start_datetime()).equals(flag) && !simpleDateFormat.format(docker.getDc_end_datetime()).equals(flag)){
+                if(docker.getDc_start_datetime().getTime() < nowDate && nowDate < docker.getDc_end_datetime().getTime()){
+                    model.addAttribute("docker",docker);
+                    return "home_shiyan/all-test";
+                }
+            }
+        }
+        model.addAttribute("docker",null);
         return "home_shiyan/all-test";
-//        return "";
     }
     //实验大厅考核模块
     @GetMapping("/kaoheModel")
-    public String kaoModelById(Model model,@RequestParam(value = "pageNum",required = true,defaultValue = "0")int pageNum,HttpSession session){
+    public String kaoModelById(Model model,@RequestParam(value = "pageNum",required = true,defaultValue = "0")int pageNum,HttpSession session) throws ParseException {
+        SimpleDateFormat simpleDateFormat = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss");
         Student student = (Student) SecurityUtils.getSubject().getPrincipal();
         Page<KaoHeModelStuDTO> kaohe = kaoheModelService.findKaoheModelStuDto(student.getId(),pageNum);
         model.addAttribute("k",kaohe);
         session.setAttribute("modulePageNum",pageNum);
         session.setAttribute("isAllModule",false);
+        Docker docker = dockerService.findDockerByStu_id(student.getId());
+        long nowDate = new Date().getTime();
+        String flag = "1314-06-21 00:00:00";
+        if(docker != null){
+            if(simpleDateFormat.format(docker.getDc_start_datetime()).equals(flag) && simpleDateFormat.format(docker.getDc_end_datetime()).equals(flag)){
+                model.addAttribute("docker",docker);
+                return "home_shiyan/index";
+            }else if(!simpleDateFormat.format(docker.getDc_start_datetime()).equals(flag) && simpleDateFormat.format(docker.getDc_end_datetime()).equals(flag)){
+                if(nowDate > docker.getDc_start_datetime().getTime()){
+                    model.addAttribute("docker",docker);
+                    return "home_shiyan/index";
+                }
+            }else if(simpleDateFormat.format(docker.getDc_start_datetime()).equals(flag) && !simpleDateFormat.format(docker.getDc_end_datetime()).equals(flag)){
+                if(nowDate < docker.getDc_end_datetime().getTime()){
+                    model.addAttribute("docker",docker);
+                    return "home_shiyan/index";
+                }
+            } else if(!simpleDateFormat.format(docker.getDc_start_datetime()).equals(flag) && !simpleDateFormat.format(docker.getDc_end_datetime()).equals(flag)){
+                if(docker.getDc_start_datetime().getTime() < nowDate && nowDate < docker.getDc_end_datetime().getTime()){
+                    model.addAttribute("docker",docker);
+                    return "home_shiyan/index";
+                }
+            }
+        }
+        model.addAttribute("docker",null);
         return "home_shiyan/index";
     }
 
@@ -357,8 +434,12 @@ public class ExpModelController {
 
     //首页跳转过来的模块
     @GetMapping("/home_exp/{id}")
-    public String homeExp(@PathVariable("id") int id,Model model){
+    public String homeExp(@PathVariable("id") int id,Model model) throws ParseException {
         Student student = (Student) SecurityUtils.getSubject().getPrincipal();
+        SimpleDateFormat simpleDateFormat = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss");
+        Docker docker = dockerService.findDockerByStu_id(student.getId());
+        long nowDate = new Date().getTime();
+        String flag = "1314-06-21 00:00:00";
         if(student.getClassId() != 0) {
             ClassModel classModel = clazzService.findById(student.getClassId());
             if (classModel.getClassIscurrent() == false) {
@@ -367,6 +448,28 @@ public class ExpModelController {
                     //考核模块
                     KaoHeModelStuDTO kaoHeModelStuDTO = kaoheModelService.findKaoHeModelStuDTOByStuId(student.getId(), id);
                     model.addAttribute("k", kaoHeModelStuDTO);
+                    if(docker != null){
+                        if(simpleDateFormat.format(docker.getDc_start_datetime()).equals(flag) && simpleDateFormat.format(docker.getDc_end_datetime()).equals(flag)){
+                            model.addAttribute("docker",docker);
+                            return "home_shiyan/kaohe_copy";
+                        }else if(!simpleDateFormat.format(docker.getDc_start_datetime()).equals(flag) && simpleDateFormat.format(docker.getDc_end_datetime()).equals(flag)){
+                            if(nowDate > docker.getDc_start_datetime().getTime()){
+                                model.addAttribute("docker",docker);
+                                return "home_shiyan/kaohe_copy";
+                            }
+                        }else if(simpleDateFormat.format(docker.getDc_start_datetime()).equals(flag) && !simpleDateFormat.format(docker.getDc_end_datetime()).equals(flag)){
+                            if(nowDate < docker.getDc_end_datetime().getTime()){
+                                model.addAttribute("docker",docker);
+                                return "home_shiyan/kaohe_copy";
+                            }
+                        } else if(!simpleDateFormat.format(docker.getDc_start_datetime()).equals(flag) && !simpleDateFormat.format(docker.getDc_end_datetime()).equals(flag)){
+                            if(docker.getDc_start_datetime().getTime() < nowDate && nowDate < docker.getDc_end_datetime().getTime()){
+                                model.addAttribute("docker",docker);
+                                return "home_shiyan/kaohe_copy";
+                            }
+                        }
+                    }
+                    model.addAttribute("docker",null);
                     return "home_shiyan/kaohe_copy";
                 }
             }
@@ -374,6 +477,28 @@ public class ExpModelController {
         //不具备考核资格
         ExpModel expModel = expModelService.findExpModelByID(id);
         model.addAttribute("exp",expModel);
+        if(docker != null){
+            if(simpleDateFormat.format(docker.getDc_start_datetime()).equals(flag) && simpleDateFormat.format(docker.getDc_end_datetime()).equals(flag)){
+                model.addAttribute("docker",docker);
+                return "home_shiyan/all-test_copy";
+            }else if(!simpleDateFormat.format(docker.getDc_start_datetime()).equals(flag) && simpleDateFormat.format(docker.getDc_end_datetime()).equals(flag)){
+                if(nowDate > docker.getDc_start_datetime().getTime()){
+                    model.addAttribute("docker",docker);
+                    return "home_shiyan/all-test_copy";
+                }
+            }else if(simpleDateFormat.format(docker.getDc_start_datetime()).equals(flag) && !simpleDateFormat.format(docker.getDc_end_datetime()).equals(flag)){
+                if(nowDate < docker.getDc_end_datetime().getTime()){
+                    model.addAttribute("docker",docker);
+                    return "home_shiyan/all-test_copy";
+                }
+            } else if(!simpleDateFormat.format(docker.getDc_start_datetime()).equals(flag) && !simpleDateFormat.format(docker.getDc_end_datetime()).equals(flag)){
+                if(docker.getDc_start_datetime().getTime() < nowDate && nowDate < docker.getDc_end_datetime().getTime()){
+                    model.addAttribute("docker",docker);
+                    return "home_shiyan/all-test_copy";
+                }
+            }
+        }
+        model.addAttribute("docker",null);
         return "home_shiyan/all-test_copy";
 
     }
@@ -395,9 +520,9 @@ public class ExpModelController {
 
     //中转站
     @GetMapping("/homeExpDispatcher/{id}")
-    public String homeExpDispatcher(@PathVariable("id") int id, Model model){
+    public String homeExpDispatcher(@PathVariable("id") int id, Model model,HttpSession session){
 
-        Student student = (Student) SecurityUtils.getSubject().getPrincipal();
+        Student student = (Student) session.getAttribute("student");
         //暂时做了修改，如果没有登录，跳转到登录页
         if(student == null){
             return "home_page/login";
